Conflicts:
	.vs/Rabbit/v14/.suo
	Queue/Opportunity/ReceiveOpportunity.cs
	Queue/bin/Debug/Main.dll
	Queue/bin/Debug/Main.pdb
	Queue/bin/Debug/Queue.dll
	Queue/bin/Debug/Queue.pdb
	Queue/obj/Debug/Queue.csprojResolveAssemblyReference.cache
	Queue/obj/Debug/Queue.dll
	Queue/obj/Debug/Queue.pdb
	Rabbit/bin/Main.dll
	Rabbit/bin/Main.pdb
	Rabbit/bin/Queue.dll
	Rabbit/bin/Queue.pdb
	Rabbit/bin/Rabbit.dll
	Rabbit/bin/Rabbit.pdb
	Rabbit/obj/Debug/Rabbit.csprojResolveAssemblyReference.cache
	Rabbit/obj/Debug/Rabbit.dll
	Rabbit/obj/Debug/Rabbit.pdb
	SalesForce/bin/Debug/Main.dll
	SalesForce/bin/Debug/Main.pdb
	SalesForce/obj/Debug/Main.dll
	SalesForce/obj/Debug/Main.pdb
	Scheduler/bin/Debug/Main.dll
	Scheduler/bin/Debug/Main.pdb
	Scheduler/bin/Debug/Queue.dll
	Scheduler/bin/Debug/Queue.pdb
	Scheduler/bin/Debug/Scheduler.exe
	Scheduler/bin/Debug/Scheduler.pdb
	Scheduler/obj/Debug/Scheduler.csprojResolveAssemblyReference.cache
	Scheduler/obj/Debug/Scheduler.exe
	Scheduler/obj/Debug/Scheduler.pdb

+++ SalesForce/Models/Opportunity.cs
@@ -14,7 +14,7 @@ namespace Main.Models
     {
         [Key]
         [Column("OppID")]
-        public Int64 OpportunityId { get; set; }
+        public int OpportunityId { get; set; }
 
         [Column ("SalesDataHeaderID")]
         public Int64 SalesDataId { get; set; }
